requirements.txt
Flask
Flask-Cors
requests
openai
numpy
regex
pinecone-client
beautifulsoup4
googlesearch-python